fix: BigInt serialization crash + Telegram 429 rate-limit
- json() reviver coerces BigInt->string; drop needless BigInt() wraps
  in serverClient/priceMonitor (values are already plain numbers)
- bot.js wraps outgoing sends in a paced queue (~1 msg/s); long 429
  penalties drop sends instead of stalling the orchestrator's awaited
  sends (which was freezing the signal loop)

+++ src/signals/priceMonitor.js
@@ -87,7 +87,7 @@ export async function monitorPriceAlerts() {
           mint: alert.mint,
           fee: signal.feeClaim ? {
             mint: alert.mint,
-            distributed: BigInt(Math.floor(signal.feeClaim.distributedSol * 1e9)),
+            distributed: Math.floor(signal.feeClaim.distributedSol * 1e9),
             shareholders: (signal.feeClaim.shareholders || []).map(h => ({ pubkey: h.address, bps: h.bps })),
           } : null,
           signature: signal.feeClaim?.signature || null,
