+++ src/signals/serverClient.js
@@ -130,7 +130,7 @@ export async function fetchServerSignals() {
       if (signal.feeClaim) {
         fee = {
           mint,
-          distributed: BigInt(Math.floor(signal.feeClaim.distributedSol * 1e9)),
+          distributed: Math.floor(signal.feeClaim.distributedSol * 1e9),
           shareholders: (signal.feeClaim.shareholders || []).map(h => ({
             pubkey: h.address,
             bps: h.bps,
